client updated project

+++ project.cpp
@@ -479,7 +479,7 @@ void onFrame(uint32_t width, uint32_t height) {
 			glm::vec3 flow;
 			fluid.velocities.front().read_interp(o.location, &flow.x);
 			
-			float creature_speed = 4.f*alice.dt;
+			float creature_speed = 4.f*(float)alice.dt;
 			glm::vec3 push = quat_uf(o.orientation) * creature_speed;
 			fluid.velocities.front().add(o.location, &push.x);
